Removing namespaces, cleaning code.

+++ index.php
@@ -1,11 +1,10 @@
 <?php
-
 require 'vendor/autoload.php';
+
 use \PhpOffice\PhpWord\PhpWord;
 use \PhpOffice\PhpWord\Template;
-use Sabre\XML;
-echo '<pre>';
-$fileName = 'file3.docx';
+
+$fileName = 'file.docx';
 
 $tagsWanted = array(
     'NUNBER_OF_FTE',
@@ -14,14 +13,20 @@ $tagsWanted = array(
     'FSSC_Company_Number'
 );
 
-$word = \PhpOffice\PhpWord\IOFactory::load($fileName);
-$tags = $word->getTags();
-foreach ($tags as $tag) {
-    if (in_array($tag['property']['tag'], $tagsWanted)) {
-        echo $tag['property']['alias'].': '.$tag['content']['content'].PHP_EOL;
+if (file_exists($fileName)) {
+
+    $word = \PhpOffice\PhpWord\IOFactory::load($fileName);
+    $tags = $word->getTags();
+    foreach ($tags as $tag) {
+        if (in_array($tag['property']['tag'], $tagsWanted)) {
+            echo $tag['property']['alias'] . ': ' . $tag['content']['content'] . PHP_EOL;
+        }
     }
+
+    // Extracts document.xml
+    $template = new Template($fileName);
+    file_put_contents(basename($fileName) . '.xml', $template->documentXML);
+} else {
+    echo "File doens't exists $fileName";
 }
-// Extracts document.xml
-$template = new Template($fileName);
-file_put_contents(basename($fileName).'.xml', $template->documentXML);
 exit;
